+++ mbuf.h
@@ -101,7 +101,7 @@ void mbuf_append_u8(mbuf_t *m, uint8_t u8);
 
 void mbuf_append_u16_be(mbuf_t *m, uint16_t u16);
 
-void mbuf_append_u32_be(mbuf_t *m, uint16_t u32);
+void mbuf_append_u32_be(mbuf_t *m, uint32_t u32);
 
 void mbuf_dump_raw_stderr(mbuf_t *m);
